refactor: separate student interview components
