+++ client/src/modules/Interview/Student/index.ts
@@ -1,0 +1,2 @@
+export { NoInterviewsAlert } from './components/NoInterviewsAlert';
+export { InterviewCard } from './components/InterviewCard';
